requirements.txt
dpckan==0.1.11
jsonpatch==1.32
psycopg2==2.9.1
PyMySQL==1.0.2
ipdb==0.13.9
ruamel.yaml==0.17.17
git+https://github.com/transparencia-mg/dtamg-py.git@main#egg=dtamg-py
